+++ framework-orm/framework-orm-mongo/src/main/java/com/rnkrsoft/framework/orm/mongo/MongoPrimaryKeyHelper.java
@@ -53,7 +53,7 @@ public class MongoPrimaryKeyHelper {
             final SequenceService sequenceService = ssc.getSequenceService(tableMetadata.getTableName());
             Object primaryKey = null;
             if (primaryKeyMetadata.getJavaType() == String.class) {
-                Integer seqNo = null;
+                Long seqNo = null;
                 try {
                     seqNo = sequenceService.nextval(schema, tableMetadata.getPrefix(), tableName, seqFeature);
                 } catch (Exception e) {
@@ -67,13 +67,13 @@ public class MongoPrimaryKeyHelper {
                 }
                 if (feature == null) {
                     if (primaryKeyMetadata.getJavaType() == String.class) {
-                        primaryKey = StringUtils.fill(Integer.toString(seqNo), true, '0', 5);
+                        primaryKey = StringUtils.fill(Long.toString(seqNo), true, '0', 5);
                     } else {
                         primaryKey = seqNo;
                     }
 
                 } else if (feature.equals(PrimaryKeyFeatureConstant.YYYY_MM_DD_HH_MM_SS_SSS_SEQUEUE5)) {
-                    primaryKey = seqFeature + StringUtils.fill(Integer.toString(seqNo), true, '0', 5);
+                    primaryKey = seqFeature + StringUtils.fill(Long.toString(seqNo), true, '0', 5);
                 }
             } else if (primaryKeyMetadata.getJavaType() == Integer.class || primaryKeyMetadata.getJavaType() == Integer.TYPE) {
                 try {
